🚧 refactor messages script into multiple files
